Adjust startup so that PIOS_Board_Init is called with the scheduler running.  This allows for the use of the RTOS facilities during initialisation.
Modify the heap code to recognise the init stack being freed by task cleanup and add it to the end of the heap.

Teach the heap how to claim all memory from the end of the BSS to the base of the init stack, whilst still enforcing the minimum size defined in the configuration file.