qt3d: enable library build and disable examples, demos, tutorials
Note that building of this version will INSTALL Qt3D headers and
libraries to the Qt installation path. Then it is possible to use
Qt3D in any applications using this QtSDK. But this does not look
like clean solution for 3rd party sources. It is supposed to be
changed so all libs will be kept withing the project build tree.