i2c-vm: Add a virtual machine that can drive i2c bus transactions
This virtual machine can be programmed by the GCS using a custom
byte code.  The optional GenericI2CSensor module will load load
the program from a settings object and will boot the I2C VM to
run the user's program.

Some of the registers from the VM are exported back to the GCS
via the GenericI2CSensor UAVO whenever the user's program requests
it.