Complete refactoring of attitude module. Added new attitude filter, that builds on the existing complementary filter approach.
This new filter is capable of using GPS and magnetometer data, although these are currently untested. Furthermore, GPS module support for velocity object is not currently enabled.

Memory usage is currently high due to the addition of two UAVO. These UAVO are for temporary development, so should be removed in the final version.