PIOS SOFTUSART: Add a method to enable or disable the soft usart functionality.
This method doesn't reconfigure any hardware but causes it to ignore the driver
to ignore the interrupt.

In addition a (priviledged) method was added to PIOS_COM to get the lower driver
id from a PIOS_COM id.  This is to allow the lower drivers (e.g. PIOS_SOFTUSART)
to be called with the main id.  The lower driver is responsible for validating the
returned id.