Added attitude direct control mode
Attidude direct control mode is a regular pid control loop
controlling a channel output to achieve an attitude. It works
just like the Attutude mode except it is not a cascade controller.
This is beneficial for controlling vehicles with reduced control over
roll/pitch/yaw so that the cascade loop no longer makes sense (ie become
impossible to tune)