Upgrade toolchain system to Qt 5.1 (Windows only, use packages for Linux and OSX)
This branch does not install separate MinGW and python since both are
included into the Qt 5.1 distribution. So it just reuses them by setting
appropriate paths and replacing some install/clean/distclean rules.

The old code is NOT removed from Makefile to show how to reintroduce
them if any further Qt distribution will be splitted again.

Qt-Creator is included (being a part of Qt 5.1 distribution) but it is
not supported by Makefile in any way. It should run, though, if launched
directly.